Reorganize and bring sanity to our lives
This breaks things down into categories in the same way we have in AOSP
settings. Easier for the user to view mods but organization is always a
good thing!

This also brings us closer to the AOSP base in where there was only one
SettingsActivity instead of two. No offense to Amir but this didn't make
a lot of sense and confused a lot of people.

Vectors pulled from 'materialdesignicons.com' so thanks to those folks!

Change-Id: Id34776cc374c9181e482dbd66c2492a5d1e36cd5